# LSP Optimization: Remove `ParseResult` and write to session types directly