--- a/sway-lsp/benches/lsp_benchmarks/compile.rs
+++ b/sway-lsp/benches/lsp_benchmarks/compile.rs
@@ -1,3 +1,5 @@
+use std::sync::Arc;
+
 use criterion::{black_box, criterion_group, Criterion};
 use lsp_types::Url;
 use sway_core::Engines;
@@ -18,8 +20,10 @@ fn benchmarks(c: &mut Criterion) {
     c.bench_function("traverse", |b| {
         let engines = Engines::default();
         let results = black_box(session::compile(&uri, &engines, None).unwrap());
+        let session = Arc::new(session::Session::new());
         b.iter(|| {
-            let _ = black_box(session::traverse(results.clone(), &engines).unwrap());
+            let _ =
+                black_box(session::traverse(results.clone(), &engines, session.clone()).unwrap());
         })
     });