--- a/sway-lsp/benches/lsp_benchmarks/mod.rs
+++ b/sway-lsp/benches/lsp_benchmarks/mod.rs
@@ -4,18 +4,16 @@ pub mod token_map;
 
 use lsp_types::Url;
 use std::{path::PathBuf, sync::Arc};
-use sway_lsp::core::session::{self, ParseResult, Session};
+use sway_lsp::core::session::{self, Session};
 
 pub async fn compile_test_project() -> (Url, Arc<Session>) {
-    let session = Session::new();
+    let session = Arc::new(Session::new());
     // Load the test project
     let uri = Url::from_file_path(benchmark_dir().join("src/main.sw")).unwrap();
     session.handle_open_file(&uri).await;
-    // Compile the project and write the parse result to the session
-    let mut parse_result = ParseResult::default();
-    session::parse_project(&uri, &session.engines.read(), None, &mut parse_result).unwrap();
-    session.write_parse_result(&mut parse_result);
-    (uri, Arc::new(session))
+    // Compile the project
+    session::parse_project(&uri, &session.engines.read(), None, session.clone()).unwrap();
+    (uri, session)
 }
 
 pub fn sway_workspace_dir() -> PathBuf {