## Description

We no longer write TokenMap or Metrics into an intermediate copy and then back to the original in `write_parse_result`. Instead, we write to them directly during traversal if the compilation was successful. 

related to #5445 

| Metric   | Before    | After     | Change (%)       |
|----------|-----------|-----------|------------------|
| traverse & write | 122.018ms | 74.353ms  | -39.07%          |

## Checklist
- [x] I have linked to any relevant issues.
- [x] I have commented my code, particularly in hard-to-understand areas.
- [ ] I have updated the documentation where relevant (API docs, the reference, and the Sway book).
- [ ] I have added tests that prove my fix is effective or that my feature works.
- [ ] I have added (or requested a maintainer to add) the necessary `Breaking*` or `New Feature` labels where relevant.
- [x] I have done my best to ensure that my PR adheres to [the Fuel Labs Code Review Standards](https://github.com/FuelLabs/rfcs/blob/master/text/code-standards/external-contributors.md).
- [x] I have requested a review from the relevant team or maintainers.